+++ packages/runtime/adapters/readmodel-adapters/readmodel-mysql/src/connect.ts
@@ -13,38 +13,48 @@ const inlineLedgerRunQuery: (
   querySQL,
   passthroughRuntimeErrors = false
 ) => {
-  let rows = null
-  for (;;) {
-    try {
-      void ([rows] = await pool.connection.query(querySQL))
-      break
-    } catch (error) {
-      if (pool.activePassthrough) {
-        if (
-          pool.PassthroughError.isPassthroughError(
-            error,
-            !!passthroughRuntimeErrors
-          )
-        ) {
-          throw new pool.PassthroughError()
+  const connection = pool.distinctMode
+    ? await pool.getConnection()
+    : pool.connection
+
+  try {
+    let rows = null
+    for (;;) {
+      try {
+        void ([rows] = await connection.query(querySQL))
+        break
+      } catch (error) {
+        if (pool.activePassthrough) {
+          if (
+            pool.PassthroughError.isPassthroughError(
+              error,
+              !!passthroughRuntimeErrors
+            )
+          ) {
+            throw new pool.PassthroughError()
+          } else {
+            throw error
+          }
         } else {
-          throw error
-        }
-      } else {
-        if (pool.PassthroughError.isPassthroughError(error, false)) {
-          await new Promise((resolve) => setTimeout(resolve, 100))
-        } else {
-          throw error
+          if (pool.PassthroughError.isPassthroughError(error, false)) {
+            await new Promise((resolve) => setTimeout(resolve, 100))
+          } else {
+            throw error
+          }
         }
       }
     }
-  }
 
-  if (rows != null && Array.isArray(rows)) {
-    rows = JSON.parse(JSON.stringify(rows))
-  }
+    if (rows != null && Array.isArray(rows)) {
+      rows = JSON.parse(JSON.stringify(rows))
+    }
 
-  return rows as Array<object>
+    return rows as Array<object>
+  } finally {
+    if (pool.distinctMode) {
+      await connection.end()
+    }
+  }
 }
 
 const makeNestedPath: MakeNestedPathMethod = (nestedPath) => {
@@ -73,18 +83,22 @@ const connect: CurrentConnectMethod = async (imports, pool, options) => {
     tablePrefix = ''
   }
 
-  const connection = await imports.MySQL.createConnection({
-    ...connectionOptions,
-    multipleStatements: true,
-  })
+  const getConnection = async () => {
+    const connection = await imports.MySQL.createConnection({
+      ...connectionOptions,
+      multipleStatements: true,
+    })
 
-  const [[{ version }]] = (await connection.query(
-    `SELECT version() AS \`version\``
-  )) as [Array<{ version: string }>, unknown]
+    const [[{ version }]] = (await connection.query(
+      `SELECT version() AS \`version\``
+    )) as [Array<{ version: string }>, unknown]
 
-  const major = +version.split('.')[0]
-  if (isNaN(major) || major < 8) {
-    throw new Error(`Supported MySQL version 8+, but got ${version}`)
+    const major = +version.split('.')[0]
+    if (isNaN(major) || major < 8) {
+      throw new Error(`Supported MySQL version 8+, but got ${version}`)
+    }
+
+    return connection
   }
 
   Object.assign(pool, {
@@ -92,10 +106,12 @@ const connect: CurrentConnectMethod = async (imports, pool, options) => {
       null,
       pool
     ) as InlineLedgerRunQueryMethod,
+    connection: await getConnection(),
+    activePassthrough: false,
+    distinctMode: false,
     tablePrefix,
     makeNestedPath,
-    connection,
-    activePassthrough: false,
+    getConnection,
     ...imports,
   })
 }