+++ packages/runtime/adapters/readmodel-adapters/readmodel-postgresql/src/find-one.ts
@@ -8,6 +8,7 @@ const findOne: CurrentStoreApi['findOne'] = async (
     tablePrefix,
     searchToWhereExpression,
     makeNestedPath,
+    splitNestedPath,
     convertResultRow,
     schemaName,
   },
@@ -20,7 +21,8 @@ const findOne: CurrentStoreApi['findOne'] = async (
     searchExpression,
     escapeId,
     escapeStr,
-    makeNestedPath
+    makeNestedPath,
+    splitNestedPath
   )
 
   const inlineSearchExpr =