+++ packages/runtime/adapters/replicators/replicator-via-api-handler/src/drop-read-model.ts
@@ -1,0 +1,10 @@
+import type { DropReadModelMethod } from './types'
+import fetch from 'node-fetch'
+
+const dropReadModel: DropReadModelMethod = async (pool, readModelName) => {
+  await fetch(`${pool.targetApplicationUrl}/api/reset-replication`, {
+    method: 'POST',
+  })
+}
+
+export default dropReadModel