+++ packages/runtime/adapters/replicators/replicator-via-api-handler/src/resume.ts
@@ -1,0 +1,8 @@
+import type { ExternalMethods } from './types'
+
+const resume: ExternalMethods['resume'] = async (pool, readModelName, next) => {
+  await pool.setReplicationPaused(pool, false)
+  await next()
+}
+
+export default resume